Split machine and module building code out of internal_rep.rs into its own file
